Implement Below Restrictions

+++ client/js-client/index.js
@@ -1,22 +1,10 @@
-define([], () => {
-  const token = localStorage.getItem('token');
-  const url = new URL(location.href);
-  const redirect = url.searchParams.get('red');
-  if (token) {
-    fetch('/api/v1/rides', {
-      headers: new Headers({
-        'x-access-token': token,
-      }),
-      method: 'GET',
-    })
-      .then(res => res.json())
-      .then(data => data.rides.length)
-      .then((isAuth) => {
-        if (isAuth && !redirect) {
-          location.href = 'user-dashboard.html';
-        } else localStorage.clear();
-      });
-  } else {
-    localStorage.clear();
-  }
+define(['./common'], (common) => {
+  const login = document
+    .getElementById('js-login');
+  login.addEventListener('click', common.redirect);
+
+  // param used by internal pages
+  const clearStorage = (new URL(location.href))
+    .searchParams.get('clear');
+  if (clearStorage) localStorage.clear();
 });
